Add support for Staatsblad versienummers (breaking changes)
There is no automatic upgrade path for Staatsblad objects of
'Verbeterblad' type. If you already have such objects in your database,
this change can result in getting double objects in your database for
each Verbeterblad that is already in your database prior to this change.
The simplest solution is to delete all existing Verbeterblad objects and
re-crawl.

Closes #15