[BREAKING CHANGES] Add support for pre-2011/12 Handeling items
This commit contains an almost complete rewrite of the Handeling
crawler. Handeling items are now retrieved using the KOOP SRU API and
parallelization of this crawling is now supported with celery.

As part of this change, the Vergadering model is removed; the
information that was stored by this model is now directly stored
in the Handeling model. Because of this, it is advised to drop all
objects of the Handeling model and rebuild your database after upgrading
to this (or any later) commit.

Closes #13